requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.2
anthropic==0.7.8 